+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Duplicate explicit names are now supported for most types (closes #4).
+- Duplicate column names for the `pandas.DataFrame` translatable type are no supported.
 
 ## [0.3.1] - 2023-03-19